Change the way to calculate the native heap size
Previously the native heap size returned by android.os.Debug's
getNativeHeapSize() is mallinfo.usmblks, which is the maximum
total allocated space. But this value is greater than current
total if trimming has occurred. In the output of 'dumpsys meminfo',
as a result, the native heap size dose not equal to the sum of
the allocated native heap size and the free native heap size.

I think this is misleading because actual and current footprint
of the native heap is usually smaller than the max footprint,
a peak number of bytes allocated to the heap.

This commit makes the native heap size returned by
getNativeHeapSize() to be the current footprint, which is the sum
of the allocated and free native heap size.

Change-Id: Ie5eec4ac43b82c8f8d7dcf9c6d7785fd9b4526eb
Signed-off-by: Kyungmin Lee <snailee@gmail.com>